update with fix

+++ src/com/hdfclife/strategy/EndowmentPremiumStrategy.java
@@ -2,7 +2,7 @@ package com.hdfclife.strategy;
 
 public class EndowmentPremiumStrategy implements PremiumStrategy{
     @Override
-    public int calculatePremium(int premium) {
+    public double calculatePremium(double premium) {
         return (premium * 108) / 100;
     }
 }
