+++ src/com/hdfclife/strategy/PremiumStrategy.java
@@ -3,5 +3,5 @@ package com.hdfclife.strategy;
 import com.hdfclife.model.Policy;
 
 public interface PremiumStrategy {
-    public int calculatePremium(int premium);
+    public double calculatePremium(double premium);
 }
